requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.11.0
astroid==4.0.2
attrs==25.4.0
audioop-lts==0.2.2
blinker==1.9.0
click==8.3.1
dill==0.4.0
discord.py==2.6.4
ffmpeg-python==0.2.0
Flask==3.1.2
frozenlist==1.8.0
future==1.0.0
gunicorn==23.0.0
idna==3.11
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
multidict==6.7.0
packaging==25.0
platformdirs==4.5.1
propcache==0.4.1
ruff==0.14.8
tomlkit==0.13.3
Werkzeug==3.1.3
yarl==1.22.0
